feat: set security release if changelog has CVE numbers
PPA releases always land in the 'Release' pocket of the PPA. Therefore, determining if a release is a security release only by the pocket it's on is not enough.

For that reason, we check the changelog for CVE numbers regardless of the security flag value and make a final determination of the release is a security release.